chore: remove redundant type exports
