+++ packages/orbit-components/src/Alert/AlertButton/types.ts
@@ -21,6 +21,3 @@ type OmittedButtonCommonProps = Omit<ButtonCommonProps, "size">;
 export type Props = {
   readonly type?: Type;
 } & OmittedButtonCommonProps;
-
-declare const AlertButton: React.FunctionComponent<Props>;
-export { AlertButton, AlertButton as default };
