+++ packages/orbit-components/src/FeatureIcon/types.ts
@@ -9,6 +9,3 @@ export interface Props extends Common.Globals {
   readonly name: Name;
   readonly alt?: string;
 }
-
-declare const FeatureIcon: React.FunctionComponent<Props>;
-export { FeatureIcon, FeatureIcon as default };
