+++ packages/orbit-components/src/Pagination/types.ts
@@ -1,8 +1,6 @@
 // Type definitions for @kiwicom/orbit-components
 // Project: http://github.com/kiwicom/orbit
 
-import * as React from "react";
-
 import * as Common from "../common/types";
 
 export type OnPageChange = (page: number) => void;
