+++ packages/orbit-components/src/Separator/types.ts
@@ -1,9 +1,6 @@
 // Type definitions for @kiwicom/orbit-components
 // Project: http://github.com/kiwicom/orbit
 
-import * as React from "react";
-import { StyledComponent } from "styled-components";
-
 import * as Common from "../common/types";
 
 export type Indent = "none" | "small" | "medium" | "large" | "XLarge" | "XXLarge";
