+++ packages/orbit-components/src/hooks/useBoundingRect/index.ts
@@ -1,6 +1,17 @@
 import * as React from "react";
 
-import { Dimensions } from "./types";
+type inexactNumber = number | null;
+
+export type Dimensions = {
+  x: inexactNumber;
+  y: inexactNumber;
+  width: inexactNumber;
+  height: inexactNumber;
+  top: inexactNumber;
+  right: inexactNumber;
+  bottom: inexactNumber;
+  left: inexactNumber;
+};
 
 const useBoundingRect = <T extends HTMLElement>(
   initialValue: Partial<Dimensions> | undefined | null,
