--- packages/orbit-components/src/hooks/useBoundingRect/types.ts
@@ -1,12 +1,0 @@
-type inexactNumber = number | null;
-
-export type Dimensions = {
-  x: inexactNumber;
-  y: inexactNumber;
-  width: inexactNumber;
-  height: inexactNumber;
-  top: inexactNumber;
-  right: inexactNumber;
-  bottom: inexactNumber;
-  left: inexactNumber;
-};
