+++ packages/orbit-components/src/hooks/useFocusTrap/index.ts
@@ -2,12 +2,13 @@ import * as React from "react";
 
 import FOCUSABLE_ELEMENT_SELECTORS from "./consts";
 import KEY_CODE_MAP from "../../common/keyMaps";
-import UseFocusTrap from "./types";
 
 interface FocusElements {
   first: HTMLElement | null;
   last: HTMLElement | null;
 }
+
+type UseFocusTrap = <T extends HTMLElement>(ref: React.RefObject<T>) => void;
 
 const manageFocus = (ref, triggered): FocusElements => {
   if (triggered && ref.current) {
@@ -24,7 +25,7 @@ const manageFocus = (ref, triggered): FocusElements => {
   return { first: null, last: null };
 };
 
-const useFocusTrap: typeof UseFocusTrap = ref => {
+const useFocusTrap: UseFocusTrap = ref => {
   const [triggered, setTriggered] = React.useState(false);
 
   React.useEffect(() => {
