--- packages/orbit-components/src/hooks/useFocusTrap/types.ts
@@ -1,5 +1,0 @@
-import React from "react";
-
-declare const useFocusTrap: <T extends HTMLElement>(ref: React.RefObject<T>) => void;
-
-export { useFocusTrap, useFocusTrap as default };
