--- packages/orbit-components/src/hooks/useIntersect/types.ts
@@ -1,9 +1,0 @@
-declare const useIntersect: (
-  intersect?: IntersectionObserverInit,
-) => {
-  ref: (el: Element | null) => void;
-  entry: IntersectionObserverEntry | null;
-  observer: IntersectionObserver | null;
-};
-
-export { useIntersect, useIntersect as default };
