+++ packages/orbit-components/src/hooks/useIsMountedRef/index.ts
@@ -1,6 +1,6 @@
 import { useRef, useEffect } from "react";
 
-const useIsMountedRef = () => {
+const useIsMountedRef = (): React.RefObject<boolean | null> => {
   const isMountedRef = useRef<boolean | null>(null);
 
   useEffect(() => {
