+++ packages/orbit-components/src/primitives/IllustrationPrimitive/types.ts
@@ -1,8 +1,6 @@
 // Type definitions for @kiwicom/orbit-components
 // Project: http://github.com/kiwicom/orbit
 
-import React from "react";
-
 import * as Common from "../../common/types";
 
 export type Size = "extraSmall" | "small" | "medium" | "large" | "display";
